Further DRY refactor on schema mapper
