+++ dozer-ingestion/src/connectors/object_store/mod.rs
@@ -1,3 +1,4 @@
+mod adapters;
 pub mod connector;
 mod executor;
 mod helper;
