+++ dozer-ingestion/src/connectors/object_store/table_reader.rs
@@ -1,4 +1,5 @@
-use crate::connectors::object_store::helper::{get_details, get_table, map_listing_options};
+use crate::connectors::object_store::adapters::DozerObjectStore;
+use crate::connectors::object_store::helper::map_listing_options;
 use crate::connectors::object_store::schema_helper::map_value_to_dozer_field;
 use crate::connectors::{ColumnInfo, TableInfo};
 use crate::errors::ObjectStoreConnectorError::TableReaderError;
@@ -12,12 +13,9 @@ use datafusion::datasource::listing::{
     ListingOptions, ListingTable, ListingTableConfig, ListingTableUrl,
 };
 use datafusion::prelude::SessionContext;
-use dozer_types::ingestion_types::{IngestionMessage, LocalStorage, S3Storage, Table};
+use dozer_types::ingestion_types::IngestionMessage;
 use dozer_types::types::{Operation, Record, SchemaIdentifier};
 use futures::StreamExt;
-use object_store::aws::{AmazonS3, AmazonS3Builder};
-use object_store::local::LocalFileSystem;
-use object_store::ObjectStore;
 use std::sync::Arc;
 use tokio::runtime::Runtime;
 
@@ -110,23 +108,6 @@ impl<T: Clone + Send + Sync> TableReader<T> {
     }
 }
 
-pub trait DozerObjectStore: Clone + Send + Sync {
-    type ObjectStore: ObjectStore;
-
-    fn table_params<'a>(
-        &'a self,
-        table_name: &str,
-    ) -> Result<DozerObjectStoreParams<'a, Self::ObjectStore>, ConnectorError>;
-}
-
-pub struct DozerObjectStoreParams<'a, T: ObjectStore> {
-    pub scheme: &'static str,
-    pub host: &'a str,
-    pub object_store: T,
-    pub table_path: String,
-    pub data_fusion_table: &'a Table,
-}
-
 pub trait Reader<T> {
     fn read_tables(&self, tables: &[TableInfo], ingestor: &Ingestor) -> Result<(), ConnectorError>;
 }
@@ -163,56 +144,5 @@ impl<T: DozerObjectStore> Reader<T> for TableReader<T> {
         }
 
         Ok(())
-    }
-}
-
-impl DozerObjectStore for S3Storage {
-    type ObjectStore = AmazonS3;
-
-    fn table_params(
-        &self,
-        table_name: &str,
-    ) -> Result<DozerObjectStoreParams<Self::ObjectStore>, ConnectorError> {
-        let table = get_table(&self.tables, table_name)?;
-        let details = get_details(&self.details)?;
-
-        let object_store = AmazonS3Builder::new()
-            .with_bucket_name(&details.bucket_name)
-            .with_region(&details.region)
-            .with_access_key_id(&details.access_key_id)
-            .with_secret_access_key(&details.secret_access_key)
-            .build()
-            .map_err(|_| ConnectorError::InitializationError)?;
-
-        Ok(DozerObjectStoreParams {
-            scheme: "s3",
-            host: &details.bucket_name,
-            object_store,
-            table_path: format!("s3://{}/{}/", details.bucket_name, table.prefix),
-            data_fusion_table: table,
-        })
-    }
-}
-
-impl DozerObjectStore for LocalStorage {
-    type ObjectStore = LocalFileSystem;
-
-    fn table_params(
-        &self,
-        table_name: &str,
-    ) -> Result<DozerObjectStoreParams<Self::ObjectStore>, ConnectorError> {
-        let table = get_table(&self.tables, table_name)?;
-        let path = get_details(&self.details)?.path.as_str();
-
-        let object_store = LocalFileSystem::new_with_prefix(path)
-            .map_err(|_| ConnectorError::InitializationError)?;
-
-        Ok(DozerObjectStoreParams {
-            scheme: "local",
-            host: path,
-            object_store,
-            table_path: format!("s3://{path}/{}/", table.prefix),
-            data_fusion_table: table,
-        })
     }
 }
